refactor(parser): split primary/string.rs into cohesive submodules (§7-8)
Split the 2805-line parser/primary/string.rs (over the 500-line limit)
into a facade plus ten themed submodules under src/parser/primary/string/,
all under 500 lines. string.rs is free parser functions addressed as
string::<name> (22 call sites); the facade re-exports each submodule
preserving its original visibility, so no external caller changes.
primary/mod.rs is untouched.

- string.rs (50)          - facade: use lines + re-exports
- interp_var.rs (484)     - try_interpolate_var (single indivisible fn)
- quoted.rs (384)         - single/double/corner/smart quoted strings
- interp_helpers.rs (376) - interp call/method/index/double-sigil helpers
- helpers.rs (359)        - perror/word/bracket/delimiter helpers
- q_string.rs (354)       - big_q_string / q_string + q-content helpers
- heredoc.rs (313)        - heredoc parsing + dedent
- escapes.rs (196)        - process_escape_sequence
- quotewords.rs (162)     - parse_quotewords_* family
- interp_content.rs (153) - finalize/interpolate string content
- qx.rs (76)              - qx_string / backtick_qx_string

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility; cross-submodule private helpers bumped to
pub(crate). Body refs to primary-level siblings rewritten to absolute
crate::parser paths. The fn multiset is byte-identical. No behavior change.
